feat: refactored the project
