+++ src/components/ui/button.tsx
@@ -1,5 +1,13 @@
 import React from "react";
-import { ButtonProps } from "../../types/ui";
+
+interface ButtonProps {
+    children: React.ReactNode;
+    onClick?: () => void;
+    variant?: "primary" | "secondary" | "danger" | "success";
+    size?: "sm" | "md" | "lg";
+    disabled?: boolean;
+    className?: string;
+}
 
 interface ExtendedButtonProps extends Omit<ButtonProps, "variant"> {
     isIconOnly?: boolean;
