+++ src/types/shopping-list.ts
@@ -3,7 +3,3 @@ export interface ShoppingItem {
     name: string;
     amount: number;
 }
-
-export interface AddNewProductProps {
-    onAdd?: (productName: string, productAmount: number) => void;
-}
